requirements.txt
Flask
pandas
faiss-cpu
numpy
openai
google-cloud-storage
selenium
requests
packaging
google-cloud-logging
filetype